+++ Assets/MeshUtility/Editor/ExportDialog.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d7ee61ea6eb4ff541831cba9ce76371f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
