+++ Assets/MeshUtility/Editor/ExportDialog/ExportDialogState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 77f1f7ea6f2c8ef4b8f284d9aa0187a4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
