+++ Assets/MeshUtility/Editor/ExportDialog/M17N.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3cde586e47f32424a82a4af6c2ce031e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
