+++ Assets/MeshUtility/Editor/ExportDialog/TabBar.cs
@@ -3,13 +3,25 @@ using System.Linq;
 using UnityEditor;
 using UnityEngine;
 
-namespace VRM
+namespace MeshUtility
 {
 
-    static class TabBar
+    public static class TabBar
     {
-        public static T OnGUI<T>(T t, GUIStyle buttonStyle, GUI.ToolbarButtonSize buttonSize) where T : Enum
+        /// <summary>
+        /// GUI.ToolbarButtonSize.FitToContentsも設定できる
+        /// </summary>
+        /// <param name="t"></param>
+        /// <param name="buttonStyle"></param>
+        /// <param name="buttonSize"></param>
+        /// <typeparam name="T"></typeparam>
+        /// <returns></returns>
+        public static T OnGUI<T>(T t, GUIStyle buttonStyle = null, GUI.ToolbarButtonSize buttonSize = GUI.ToolbarButtonSize.Fixed) where T : Enum
         {
+            if (buttonStyle == null)
+            {
+                buttonStyle = "LargeButton";
+            }
             using (new EditorGUILayout.HorizontalScope())
             {
                 GUILayout.FlexibleSpace();
@@ -19,6 +31,7 @@ namespace VRM
                 return (T)(object)value;
             }
         }
+
         static class TabCache<T> where T : Enum
         {
             private static GUIContent[] _tabToggles = null;
