+++ Assets/MeshUtility/Editor/ExportDialog/TabBar.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 639b5c2a1cd8f5345b7d8a799712e25f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
