+++ Assets/MeshUtility/Editor/ExportDialog/Valildators.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1c450c4a7f888424490abbbc29c10a1a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
