+++ Assets/MeshUtility/Editor/ExportDialog/Valildators/HumanoidValidator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9013f2150f3c6d943a6f6b6394f7a1ff
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
