+++ Assets/MeshUtility/Editor/ExportDialog/Valildators/MaterialValidator.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 515984b15b5b5bf4baca4fa2d6344fcf
+guid: aa8f593f0ade99a4d9cc23947cdded6e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
