+++ Assets/MeshUtility/Editor/ExportDialog/Valildators/NameValidator.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7829d135d35830f4bb9235eb10b3de1f
+guid: 27720b3b18437104b91e44c3f66166ee
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
