+++ Assets/MeshUtility/Runtime/MeshExportInfo.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 84a3f0f4ae5319a46839f08a76ae0edd
+guid: 2af6d95e4a0e06944b34d891d798b88b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
