+++ Assets/MeshUtility/Runtime/MeshUtility.asmdef
@@ -1,3 +1,15 @@
-﻿{
-	"name": "MeshUtility"
+{
+    "name": "MeshUtility",
+    "references": [
+        "UniUnlit",
+        "ShaderProperty.Runtime"
+    ],
+    "optionalUnityReferences": [],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": []
 }
