+++ Assets/VRM/UniVRM/Editor/FirstPerson/VRMFirstPersonEditor.cs
@@ -1,4 +1,5 @@
-﻿using UnityEditor;
+﻿using MeshUtility;
+using UnityEditor;
 using UnityEngine;
 
 namespace VRM
