+++ Assets/VRM/UniVRM/Editor/Format/VRMExportMeshes.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using MeshUtility;
 using UnityEngine;
 
 namespace VRM
@@ -55,14 +56,14 @@ namespace VRM
             return false;
         }
 
-        public List<UniGLTF.MeshExportInfo> Meshes = new List<UniGLTF.MeshExportInfo>();
+        public List<MeshUtility.MeshExportInfo> Meshes = new List<MeshUtility.MeshExportInfo>();
 
         public int ExpectedExportByteSize => Meshes.Where(x => x.IsRendererActive).Sum(x => x.ExportByteSize);
 
         List<Validation> m_validations = new List<Validation>();
         public IEnumerable<Validation> Validations => m_validations;
 
-        public static void CalcMeshSize(ref UniGLTF.MeshExportInfo info,
+        public static void CalcMeshSize(ref MeshUtility.MeshExportInfo info,
                                         string relativePath, VRMExportSettings settings, IReadOnlyList<BlendShapeClip> clips)
         {
             var sb = new StringBuilder();
@@ -132,11 +133,11 @@ namespace VRM
             sb.Append($") x {info.Mesh.vertexCount}");
             switch (info.VertexColor)
             {
-                case UniGLTF.MeshExportInfo.VertexColorState.ExistsAndIsUsed:
-                case UniGLTF.MeshExportInfo.VertexColorState.ExistsAndMixed: // エクスポートする
+                case MeshUtility.MeshExportInfo.VertexColorState.ExistsAndIsUsed:
+                case MeshUtility.MeshExportInfo.VertexColorState.ExistsAndMixed: // エクスポートする
                     sb.Insert(0, "[use vcolor]");
                     break;
-                case UniGLTF.MeshExportInfo.VertexColorState.ExistsButNotUsed:
+                case MeshUtility.MeshExportInfo.VertexColorState.ExistsButNotUsed:
                     sb.Insert(0, "[remove vcolor]");
                     break;
             }
@@ -150,7 +151,7 @@ namespace VRM
             info.Summary = sb.ToString();
         }
 
-        bool TryGetMeshInfo(GameObject root, Renderer renderer, IReadOnlyList<BlendShapeClip> clips, VRMExportSettings settings, out UniGLTF.MeshExportInfo info)
+        bool TryGetMeshInfo(GameObject root, Renderer renderer, IReadOnlyList<BlendShapeClip> clips, VRMExportSettings settings, out MeshUtility.MeshExportInfo info)
         {
             info = default;
             if (root == null)
@@ -186,7 +187,7 @@ namespace VRM
                 return false;
             }
 
-            info.VertexColor = UniGLTF.MeshExportInfo.DetectVertexColor(info.Mesh, info.Renderer.sharedMaterials);
+            info.VertexColor = MeshUtility.MeshExportInfo.DetectVertexColor(info.Mesh, info.Renderer.sharedMaterials);
 
             var relativePath = UniGLTF.UnityExtensions.RelativePathFrom(renderer.transform, root.transform);
             CalcMeshSize(ref info, relativePath, settings, clips);
@@ -216,7 +217,7 @@ namespace VRM
 
             foreach (var renderer in ExportRoot.GetComponentsInChildren<Renderer>(true))
             {
-                if (TryGetMeshInfo(ExportRoot, renderer, clips, settings, out UniGLTF.MeshExportInfo info))
+                if (TryGetMeshInfo(ExportRoot, renderer, clips, settings, out MeshUtility.MeshExportInfo info))
                 {
                     Meshes.Add(info);
                 }
